+++ datamodel/odata/odata-generator/src/main/java/com/sap/cloud/sdk/datamodel/odata/generator/NavigationPropertyMethodsGenerator.java
@@ -300,7 +300,7 @@ class NavigationPropertyMethodsGenerator
                             associatedEntity.name(),
                             navigationProperty.getEdmName()));
             setterMethod.javadoc().add(JavadocUtils.getLazyWarningMessage(navigationProperty, entityClass));
-            final JVar setterParam = setterMethod.param(JMod.FINAL, returnType, "value");
+            final JVar setterParam = setterMethod.param(JMod.FINAL, returnType, "cloudSdkValue");
             setterMethod
                 .javadoc()
                 .addParam(setterParam)
@@ -325,7 +325,7 @@ class NavigationPropertyMethodsGenerator
                             "Overwrites the associated <b>%s</b> entity for the loaded navigation property <b>%s</b>.",
                             associatedEntity.name(),
                             navigationProperty.getEdmName()));
-            final JVar setterParam = setterMethod.param(JMod.FINAL, returnType, "value");
+            final JVar setterParam = setterMethod.param(JMod.FINAL, returnType, "cloudSdkValue");
             setterMethod
                 .javadoc()
                 .addParam(setterParam)
@@ -559,7 +559,7 @@ class NavigationPropertyMethodsGenerator
                 .decl(
                     JMod.FINAL,
                     codeModel.ref(Object.class),
-                    "value",
+                    "cloudSdkValue",
                     JExpr
                         .direct(CommonConstants.INLINE_MAP_NAME)
                         .invoke("remove")
@@ -622,7 +622,7 @@ class NavigationPropertyMethodsGenerator
         final JType returnType = classMemberField.type();
         final JFieldVar buildField = builderClass.field(JMod.PRIVATE, returnType, classMemberName, init);
         final JMethod origMethod = builderClass.method(JMod.PRIVATE, builderClass, classMemberName);
-        final JVar origMethodVal = origMethod.param(JMod.FINAL, returnType, "value");
+        final JVar origMethodVal = origMethod.param(JMod.FINAL, returnType, "cloudSdkValue");
         if( isOneToMany ) {
             origMethod.body().invoke(buildField, "addAll").arg(origMethodVal);
         } else {
@@ -640,7 +640,7 @@ class NavigationPropertyMethodsGenerator
 
             final JInvocation invocation;
             if( isOneToMany ) {
-                final JVar fixedMethodParam = fixedMethod.varParam(associatedEntity, "value");
+                final JVar fixedMethodParam = fixedMethod.varParam(associatedEntity, "cloudSdkValue");
                 invocation =
                     JExpr
                         .invoke(origMethod)
@@ -651,7 +651,7 @@ class NavigationPropertyMethodsGenerator
                     .addParam(fixedMethodParam)
                     .add(String.format("The %ss to build this %s with.", associatedEntity.name(), entityClass.name()));
             } else {
-                final JVar fixedMethodParam = fixedMethod.param(JMod.FINAL, associatedEntity, "value");
+                final JVar fixedMethodParam = fixedMethod.param(JMod.FINAL, associatedEntity, "cloudSdkValue");
                 invocation = JExpr.invoke(origMethod).arg(fixedMethodParam);
 
                 fixedMethod
@@ -709,7 +709,7 @@ class NavigationPropertyMethodsGenerator
         final String origName = originalField.name();
         final JMethod originalMethod = builderClass.method(JMod.PUBLIC, builderClass, origName);
         final JFieldVar origBuildField = builderClass.field(JMod.PRIVATE, origType, origName, JExpr._null());
-        final JVar originalMethodParameter = originalMethod.param(JMod.FINAL, origType, "value");
+        final JVar originalMethodParameter = originalMethod.param(JMod.FINAL, origType, "cloudSdkValue");
 
         originalMethod.javadoc().add(originalField.javadoc());