+++ datamodel/odata/odata-generator/src/test/resources/oDataGeneratorIntegrationTest/groceryStore/output/testcomparison/namespaces/sdkgrocerystore/Customer.java
@@ -252,13 +252,13 @@ public class Customer
         // navigation properties
         {
             if ((cloudSdkValues).containsKey("Address")) {
-                final Object value = (cloudSdkValues).remove("Address");
-                if (value instanceof Map) {
+                final Object cloudSdkValue = (cloudSdkValues).remove("Address");
+                if (cloudSdkValue instanceof Map) {
                     if (toAddress == null) {
                         toAddress = new Address();
                     }
                     @SuppressWarnings("unchecked")
-                    final Map<String, Object> inputMap = ((Map<String, Object> ) value);
+                    final Map<String, Object> inputMap = ((Map<String, Object> ) cloudSdkValue);
                     toAddress.fromMap(inputMap);
                 }
             }
@@ -398,11 +398,11 @@ public class Customer
     /**
      * Overwrites the associated <b>Address</b> entity for the loaded navigation property <b>Address</b>.
      * 
-     * @param value
+     * @param cloudSdkValue
      *     New <b>Address</b> entity.
      */
-    public void setAddress(final Address value) {
-        toAddress = value;
+    public void setAddress(final Address cloudSdkValue) {
+        toAddress = cloudSdkValue;
     }
 
 
@@ -414,22 +414,22 @@ public class Customer
 
         private Address toAddress;
 
-        private Customer.CustomerBuilder toAddress(final Address value) {
-            toAddress = value;
+        private Customer.CustomerBuilder toAddress(final Address cloudSdkValue) {
+            toAddress = cloudSdkValue;
             return this;
         }
 
         /**
          * Navigation property <b>Address</b> for <b>Customer</b> to single <b>Address</b>.
          * 
-         * @param value
+         * @param cloudSdkValue
          *     The Address to build this Customer with.
          * @return
          *     This Builder to allow for a fluent interface.
          */
         @Nonnull
-        public Customer.CustomerBuilder address(final Address value) {
-            return toAddress(value);
+        public Customer.CustomerBuilder address(final Address cloudSdkValue) {
+            return toAddress(cloudSdkValue);
         }
 
     }